Refactored UI, removed unused views, implemented proper save & load mechanisms, and added a settings dialog
Utilized JavaFX ObservableList instead of List in TaskManager and FileHandler for better UI synchronization.

Removed legacy navigation views and associated controllers to simplify application flow.

Fixed bugs related to the save and load methods, and connected the methods to the UI allowing for seamless load and save functionality when opening and closing the app respectively.

Introduced a new settings dialog with a basic dark mode option (TO-DO: dark mode implementation)

Improved task editing and completion logic for more robust user feedback and data handling.

Upgraded JavaFX dependencies to the latest version for compatibility.